Split out getThemeHelpUrl
