+++ client/state/themes/selectors/get-theme-help-url.js
@@ -1,0 +1,29 @@
+/**
+ * Internal dependencies
+ */
+import config from 'config';
+import { getSiteSlug } from 'state/sites/selectors';
+import { oldShowcaseUrl } from 'state/themes/utils';
+
+import 'state/themes/init';
+
+/**
+ * Returns the URL for a given theme's support page.
+ *
+ * @param  {object}  state   Global state tree
+ * @param  {string}  themeId Theme ID
+ * @param  {?number} siteId  Site ID to optionally use as context
+ * @returns {?string}         Theme support page URL
+ */
+export function getThemeHelpUrl( state, themeId, siteId ) {
+	if ( ! themeId ) {
+		return null;
+	}
+
+	let baseUrl = oldShowcaseUrl + themeId;
+	if ( config.isEnabled( 'manage/themes/details' ) ) {
+		baseUrl = `/theme/${ themeId }/support`;
+	}
+
+	return baseUrl + ( siteId ? `/${ getSiteSlug( state, siteId ) }` : '' );
+}
